refactor: use middleware as security handler directly
